refactor: move DicePool into lib.rs
