+++ .gitignore
@@ -4,4 +4,4 @@ result
 
 # Added by cargo
 
-/target
+**/target
